+++ frontend/src/components/home/EnrollmentCard.tsx
