+++ frontend/src/components/home/SuggestionCard.tsx
